Change battery percentage calculation.
Compensate for chipset inaccuracies in returning battery percentage
the way Samsung did it in its Froyo rom. This allows the battery to
reach 100% (may require bump charging) and no visible battery drop
is noticable (even after a reboot). The precision is also increased
(no 2% jump) from the current method of calculation because the
calculation is done where we have extra precision information from
the chipset.

Change-Id: Id7f15dd3906f7a12949683c0d921ce7a8498e70a