Allow RGB color tuning with voodoo color interface.
The SGS's screen color unfortunately varies greatly (some phones
have very blue-ish screens) leading to the need for a user-tunable
color.

supercurio's Voodoo Color is used for the skeletal code, but the
implementation is different. Color multiplier regards U32_MAX as
100% of the default setting of a particular color and anything
less than that will be reduced as a percentage of U32_MAX.

If the user does not use the Voodoo Control App to tune color,
the user will not see any difference in color after applying this
patch.

Change-Id: If09996ae1a59c1fdc7df2dc1c45f019f61c9b3d6